Product add with image upload, delete and order invoice creation done

+++ src/contrllers/adminController.js
@@ -1,24 +1,66 @@
+const { mongoose } = require("mongoose");
+const product = require("../models/product");
 const Product = require("../models/product");
+const { validationResult } = require("express-validator");
+const fileHelper = require("../../utils/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/add-product", {
     docTitle: "Add Product",
     path: "/admin/add-product",
-    formCss: true,
-    productCss: true,
-    activeAddProduct: true,
+    product: {
+      title: "",
+      price: "",
+      description: "",
+    },
     editing: false,
+    hasError: false,
+    validationErrors: [],
+    errorMessage: null,
   });
 };
 
 exports.postAddProduct = (req, res, next) => {
-  const { title, image_url, description, price } = req.body;
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).render("admin/add-product", {
+      docTitle: "Add Product",
+      path: "/admin/add-product",
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      editing: false,
+      validationErrors: errors.array(),
+      errorMessage: errors.array()[0].msg,
+    });
+  }
+  const { title, description, price } = req.body;
+  const image = req.file;
+  if (!image) {
+    return res.status(422).render("admin/add-product", {
+      docTitle: "Add Product",
+      path: "/admin/add-product",
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      editing: false,
+      validationErrors: errors.array(),
+      errorMessage: "Attached file is not an Image.",
+    });
+  }
 
+  const imageUrl = image.path;
   const product = new Product({
     title: title,
     price: price,
     description: description,
-    imageUrl: image_url,
+    imageUrl: imageUrl,
     userId: req.user,
   });
 
@@ -28,7 +70,11 @@ exports.postAddProduct = (req, res, next) => {
       console.log("Product created");
       return res.redirect("/admin/products");
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.getEditProduct = (req, res, next) => {
@@ -50,11 +96,29 @@ exports.getEditProduct = (req, res, next) => {
         res.redirect("/");
       }
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.postEditProduct = (req, res, next) => {
-  const { title, image_url, description, price, id } = req.body;
+  const { title, description, price, id } = req.body;
+  const image = req.file;
+  Product.findOne({ _id: id, userId: req.user._id }).then((product) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(422).render("admin/edit-product", {
+        docTitle: "Edit Product",
+        path: "/admin/edit-product",
+        editing: true,
+        validationErrors: errors.array(),
+        errorMessage: errors.array()[0].msg,
+        product: product,
+      });
+    }
+  });
 
   Product.findById(id)
     .then((product) => {
@@ -64,22 +128,40 @@ exports.postEditProduct = (req, res, next) => {
       product.title = title;
       product.price = price;
       product.description = description;
-      product.imageUrl = image_url;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save().then((result) => {
         console.log("Product Updated");
         res.redirect("/admin/products");
       });
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const { id } = req.body;
-  Product.deleteOne({ _id: id, userId: req.user._id })
+  Product.findById(id)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: id, userId: req.user._id });
+    })
     .then((result) => {
       res.redirect("/admin/products");
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.getAdminProducts = (req, res, next) => {
@@ -91,5 +173,9 @@ exports.getAdminProducts = (req, res, next) => {
         path: "/admin/products",
       });
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
